Factorize cypress confs and commands
