--- centreon/cypress.d.ts
@@ -1,13 +1,0 @@
-import { MountOptions, MountReturn } from 'cypress/react';
-import { MemoryRouterProps } from 'react-router-dom';
-
-declare global {
-  namespace Cypress {
-    interface Chainable {
-      mount: (
-        component: React.ReactNode,
-        options?: MountOptions & { routerProps?: MemoryRouterProps },
-      ) => Cypress.Chainable<MountReturn>;
-    }
-  }
-}
