+++ centreon/packages/js-config/cypress/component/enableVisualTesting.ts
@@ -1,0 +1,21 @@
+import { addMatchImageSnapshotCommand } from 'cypress-image-snapshot/command';
+
+const enableVisualTesting = (cypressFolder = 'cypress'): void => {
+  if (Cypress.config('isInteractive')) {
+    Cypress.Commands.add('matchImageSnapshot', () => {
+      cy.log('Skipping snapshot');
+    });
+
+    return;
+  }
+
+  addMatchImageSnapshotCommand({
+    capture: 'viewport',
+    customDiffConfig: { threshold: 0.1 },
+    customSnapshotsDir: `${cypressFolder}/visual-testing-snapshots`,
+    failureThreshold: 0.06,
+    failureThresholdType: 'percent'
+  });
+};
+
+export default enableVisualTesting;
