+++ centreon/tests/e2e/cypress/e2e/Autologin/common.ts
@@ -1,7 +1,4 @@
-import {
-  applyConfigurationViaClapi,
-  executeActionViaClapi
-} from '../../commons';
+import { applyConfigurationViaClapi } from '../../commons';
 
 const initializeConfigACLAndGetLoginPage = (): Cypress.Chainable => {
   return cy
@@ -14,10 +11,12 @@ const initializeConfigACLAndGetLoginPage = (): Cypress.Chainable => {
 
 const removeContact = (): Cypress.Chainable => {
   return cy.setUserTokenApiV1().then(() => {
-    executeActionViaClapi({
-      action: 'DEL',
-      object: 'CONTACT',
-      values: 'user1'
+    cy.executeActionViaClapi({
+      bodyContent: {
+        action: 'DEL',
+        object: 'CONTACT',
+        values: 'user1'
+      }
     });
   });
 };
