+++ centreon/tests/e2e/cypress/e2e/Web-SSO/common.ts
@@ -1,5 +1,3 @@
-import { executeActionViaClapi } from '../../commons';
-
 const injectWebSSOScriptsIntoContainer = (): Cypress.Chainable => {
   return cy
     .exec(
@@ -26,15 +24,17 @@ const injectWebSSOScriptsIntoContainer = (): Cypress.Chainable => {
 const initializeWebSSOUserAndGetLoginPage = (): Cypress.Chainable => {
   return cy
     .fixture('resources/clapi/contact-web-sso/web-sso-authentication-user.json')
-    .then((contact) => executeActionViaClapi(contact))
+    .then((contact) => cy.executeActionViaClapi({ bodyContent: contact }));
 };
 
 const removeWebSSOContact = (): Cypress.Chainable => {
   return cy.setUserTokenApiV1().then(() => {
-    executeActionViaClapi({
-      action: 'DEL',
-      object: 'CONTACT',
-      values: 'oidc'
+    cy.executeActionViaClapi({
+      bodyContent: {
+        action: 'DEL',
+        object: 'CONTACT',
+        values: 'oidc'
+      }
     });
   });
 };
